hydra/proxy: merge pmip_pmi_v1.c and pmip_pmi_v2.c
Merge for now. We will refactor next.
