+++ src/pm/hydra/proxy/Makefile.mk
@@ -9,8 +9,7 @@ hydra_pmi_proxy_SOURCES = \
     proxy/pmip.c \
     proxy/pmip_cb.c \
     proxy/pmip_utils.c \
-    proxy/pmip_pmi_v1.c \
-    proxy/pmip_pmi_v2.c
+    proxy/pmip_pmi.c
 
 hydra_pmi_proxy_CPPFLAGS = $(AM_CPPFLAGS) -I$(top_srcdir)/proxy
 hydra_pmi_proxy_LDFLAGS = $(external_ldflags) -L$(top_builddir)
